refactor: build 성공 위한 코드 수정 (map 주석 처리, styles 파일 이동)
