--- src/components/add-store/styles.tsx
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-import { Card, Text, Spacer, Input, Button, Select } from '@geist-ui/react';
-
-export const StyledInput = styled(Input)`
-  width: 18rem;
-  margin-bottom: 4rem;
-`;
