--- src/pages/chat-list/styles.tsx
@@ -1,1 +1,0 @@
-import styled from 'styled-components';
