--- src/pages/main/styles.tsx
@@ -1,1 +1,0 @@
-import styled from 'styled-components';
